Refactor HMAC generation into Generator
Now that the code required is much simpler, there's no need for a separate file.
